Validando dashboard, se agrego middleware en backend y frontend

+++ frontend/src/services/ingrediente.js
@@ -1,11 +1,24 @@
 import store from "@/store";
+import utility from '@/utility';
 
 const ruta = store.state.apiUrl
 
 const read =  async () => {
   try {
-    const respuesta = await fetch(`${ruta}/ingrediente`)
-    return await respuesta.json();
+    const { exist, token } = utility.getToken();
+    if(exist) {
+      const respuesta = await fetch(`${ruta}/ingrediente`, {
+        headers: {
+          'Content-Type': 'application/json',
+          'Authorization': `Bearer ${token}`
+        }
+      })
+      return await respuesta.json();
+    }
+    utility.removeToken();
+    return {
+      error: 'El token es invalido',
+    };
   } catch (error) {
     console.log(error)
   }
@@ -13,14 +26,22 @@ const read =  async () => {
 
 const add =  async (datos) => {
   try {
-    const respuesta = await fetch(`${ruta}/ingrediente`, {
-      method: 'POST',
-      body: JSON.stringify(datos),
-      headers: {
-        'Content-Type': 'application/json'
-      }
-    })
-    return await respuesta.json();
+    const { exist, token } = utility.getToken();
+    if(exist) {
+      const respuesta = await fetch(`${ruta}/ingrediente`, {
+        method: 'POST',
+        body: JSON.stringify(datos),
+        headers: {
+          'Content-Type': 'application/json',
+          'Authorization': `Bearer ${token}`
+        }
+      })
+      return await respuesta.json();
+    }
+    utility.removeToken();
+    return {
+      error: 'El token es invalido',
+    };
   } catch (error) {
     console.log(error)
   }
@@ -28,14 +49,22 @@ const add =  async (datos) => {
 
 const update =  async (datos, id) => {
   try {
-    const respuesta = await fetch(`${ruta}/ingrediente/${id}`, {
-      method: 'PUT',
-      body: JSON.stringify(datos),
-      headers: {
-        'Content-Type': 'application/json'
-      }
-    })
-    return await respuesta.json();
+    const { exist, token } = utility.getToken();
+    if(exist) {
+      const respuesta = await fetch(`${ruta}/ingrediente/${id}`, {
+        method: 'PUT',
+        body: JSON.stringify(datos),
+        headers: {
+          'Content-Type': 'application/json',
+          'Authorization': `Bearer ${token}`
+        }
+      })
+      return await respuesta.json();
+    }
+    utility.removeToken();
+    return {
+      error: 'El token es invalido',
+    };
   } catch (error) {
     console.log(error)
   }
@@ -43,13 +72,21 @@ const update =  async (datos, id) => {
 
 const remove =  async (id) => {
   try {
-    const respuesta = await fetch(`${ruta}/ingrediente/${id}`, {
-      method: 'DELETE',
-      headers: {
-        'Content-Type': 'application/json'
-      }
-    })
-    return await respuesta.json();
+    const { exist, token } = utility.getToken();
+    if(exist) {
+      const respuesta = await fetch(`${ruta}/ingrediente/${id}`, {
+        method: 'DELETE',
+        headers: {
+          'Content-Type': 'application/json',
+          'Authorization': `Bearer ${token}`
+        }
+      })
+      return await respuesta.json();
+    }
+    utility.removeToken();
+    return {
+      error: 'El token es invalido',
+    };
   } catch (error) {
     console.log(error)
   }
@@ -57,8 +94,20 @@ const remove =  async (id) => {
 
 const all =  async () => {
   try {
-    const respuesta = await fetch(`${ruta}/ingrediente/all`)
-    return await respuesta.json();
+    const { exist, token } = utility.getToken();
+    if(exist) {
+      const respuesta = await fetch(`${ruta}/ingrediente/all`, {
+        headers: {
+          'Content-Type': 'application/json',
+          'Authorization': `Bearer ${token}`
+        }
+      })
+      return await respuesta.json();
+    }
+    utility.removeToken();
+    return {
+      error: 'El token es invalido',
+    };
   } catch (error) {
     console.log(error)
   }
